code/ssd/requirements.txt
torchvision==0.2.0
numpy
pillow
tensorboardX
opencv-python
plyfile